# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ghar Ka Sathi Production APIs",
    "version": "1.0.0",
    "description": "Comprehensive interactive API documentation for customer, expert, and admin application workflows. Built to support 80+ endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/bookings": {
      "post": {
        "tags": ["9. Bookings Lifecycle"],
        "summary": "Create new service booking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryId": { "type": "string", "example": "cat_plumber" },
                  "subServiceId": { "type": "string", "example": "sub_pl_leak" },
                  "dateTime": { "type": "string", "example": "2026-07-25T10:00:00Z" },
                  "timeSlot": { "type": "string", "example": "10:00 AM - 12:00 PM" },
                  "address": { "type": "string", "example": "Sector 15 Noida" },
                  "problemDescription": { "type": "string", "example": "Kitchen pipeline leaks" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" }
        }
      },
      "get": {
        "tags": ["9. Bookings Lifecycle"],
        "summary": "Get Bookings History (Filtered)",
        "parameters": [
          { "name": "filter", "in": "query", "schema": { "type": "string", "example": "Pending" } }
        ],
        "responses": {
          "200": { "description": "Bookings array" }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "tags": ["9. Bookings Lifecycle"],
        "summary": "Get Booking Details by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Details" }
        }
      }
    },
    "/bookings/{id}/cancel": {
      "patch": {
        "tags": ["9. Bookings Lifecycle"],
        "summary": "Cancel booking request",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Cancelled" }
        }
      }
    },
    "/bookings/{id}/reschedule": {
      "patch": {
        "tags": ["9. Bookings Lifecycle"],
        "summary": "Reschedule booking time slot",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dateTime": { "type": "string", "example": "2026-07-26T14:00:00Z" },
                  "timeSlot": { "type": "string", "example": "02:00 PM - 04:00 PM" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Rescheduled" }
        }
      }
    },
    "/bookings/{id}/rebook": {
      "post": {
        "tags": ["9. Bookings Lifecycle"],
        "summary": "Rebook a past booking",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "201": { "description": "New duplicate booking created" }
        }
      }
    },
    "/bookings/{id}/status": {
      "get": {
        "tags": ["10. Tracking & Live Status"],
        "summary": "Get Live Status timeline state",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Status value" }
        }
      }
    },
    "/bookings/{id}/tracking": {
      "get": {
        "tags": ["10. Tracking & Live Status"],
        "summary": "Get Live Expert Coordinates",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Latitude, longitude and ETA values" }
        }
      }
    },
    "/bookings/{id}/timeline": {
      "get": {
        "tags": ["10. Tracking & Live Status"],
        "summary": "Get Historical Status Timeline",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Timeline log events" }
        }
      }
    },
    "/bookings/{id}/expert": {
      "get": {
        "tags": ["11. Expert Partner Details"],
        "summary": "Get Assigned Expert Details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Provider stats" }
        }
      }
    },
    "/bookings/expert/profile/{id}": {
      "get": {
        "tags": ["11. Expert Partner Details"],
        "summary": "Get Expert Profile details by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
...
    "/bookings/call/log": {
      "post": {
        "tags": ["11. Expert Partner Details"],
        "summary": "Log a Voice Call",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingId": { "type": "string", "example": "bk_123" },
                  "durationSeconds": { "type": "number", "example": 45 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Call logged" }
        }
      }
    },
    "/bookings/{id}/otp": {
      "get": {
        "tags": ["12. OTP Verification"],
        "summary": "Get Booking Arrival OTP Code",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Returns 4-digit code" }
        }
      }
    },
    "/bookings/{id}/verify-otp": {
      "post": {
        "tags": ["12. OTP Verification"],
        "summary": "Verify OTP code to Start Service",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": { "type": "string", "example": "4280" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Job started status set" }
        }
      }
    },
    "/bookings/{id}/extra-work": {
      "get": {
        "tags": ["13. Additional Work / Material Approval"],
        "summary": "Check Pending Extra Work Request details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Extra cost detail values" }
        }
      },
      "post": {
        "tags": ["13. Additional Work / Material Approval"],
        "summary": "Submit Extra Work Request (Expert)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cost": { "type": "number", "example": 150 },
                  "reason": { "type": "string", "example": "Tap replacement spindle" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Extra cost proposed" }
        }
      }
    },
    "/bookings/{id}/approve-extra": {
      "post": {
        "tags": ["13. Additional Work / Material Approval"],
        "summary": "Approve Extra Work Invoice Cost",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Invoice updated and extra work approved" }
        }
      }
    },
    "/bookings/{id}/reject-extra": {
      "post": {
        "tags": ["13. Additional Work / Material Approval"],
        "summary": "Reject Extra Work Proposal",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Cost proposal rejected" }
        }
      }
    },
    "/bookings/{id}/invoice": {
      "get": {
        "tags": ["14. Invoice Details"],
        "summary": "Get Final Invoice breakdown details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Detailed bill parameters" }
        }
      }
    },
    "/bookings/{id}/invoice/pdf": {
      "get": {
        "tags": ["14. Invoice Details"],
        "summary": "Download PDF Invoice document link",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "PDF direct link" }
        }
      }
    },
...
    "/review": {
      "post": {
        "tags": ["17. Ratings & Reviews"],
        "summary": "Submit service rating feedback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingId": { "type": "string", "example": "bk_123" },
                  "stars": { "type": "number", "example": 5 },
                  "comment": { "type": "string", "example": "Great work" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Reviewed saved" }
        }
      }
    },
    "/review/{bookingId}": {
      "get": {
        "tags": ["17. Ratings & Reviews"],
        "summary": "Get Review details of booking",
        "parameters": [
          { "name": "bookingId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Review stars & details" }
        }
      }
    },
...
  }
}